# backend/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Theatre Booking API",
    "version": "1.0.0",
    "description": "Current endpoints for the Theatre Booking API",
    "contact": {
      "name": "Group 9"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://theatrebooking.onrender.com",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/bookings/{id}": {
      "get": {
        "summary": "Retrieve booking details by ID",
        "description": "Retrieve booking details by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Booking not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings": {
      "post": {
        "summary": "Book a show",
        "description": "Book a show by providing show ID, date, price, number of tickets booked, and booking details",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "showId": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "price": {
                    "type": "number"
                  },
                  "ticketsBooked": {
                    "type": "integer"
                  },
                  "bookingDetails": {
                    "$ref": "#/components/schemas/Booking"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Booking created successfully"
                    },
                    "booking": {
                      "$ref": "#/components/schemas/Booking"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Not enough tickets available or wrong price",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "failure"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not enough tickets available or wrong price"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Show not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Show not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "failure"
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/shows": {
      "get": {
        "summary": "Retrieve a list of shows",
        "description": "Retrieve a list of all shows",
        "responses": {
          "200": {
            "description": "A list of shows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Show"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/genres": {
      "get": {
        "summary": "Retrieve a list of genres",
        "description": "Retrieve a list of all genres",
        "responses": {
          "200": {
            "description": "A list of genres",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "Action"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/dates": {
      "get": {
        "summary": "Retrieve a list of dates",
        "description": "Retrieve a list of all dates",
        "responses": {
          "200": {
            "description": "A list of dates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "example": "2024-04-15"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Booking": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "showName": {
            "type": "string"
          },
          "ticketsBooked": {
            "type": "number"
          }
        }
      },
      "Show": {
        "type": "object",
        "properties": {
          "showName": {
            "type": "string"
          },
          "genre": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "briefDescription": {
            "type": "string"
          },
          "session": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string"
                },
                "ticketsAvailability": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketsAvailability"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}